Patch Role on product edit

+++ src/src/Controller/StuffController.php
@@ -76,6 +76,12 @@ class StuffController extends AbstractController
      */
     public function show(Product $product, Breadcrumbs $breadcrumbs): Response
     {
+        if ($this->getUser() !== $product->getUser())
+        {
+            $response = new Response();
+            $response->setStatusCode(403);
+            return $response;
+        }
         $breadcrumbs->addItem("Stuff", $this->get("router")->generate("stuff_index"));
         $breadcrumbs->addItem("Show", $this->get("router")->generate("stuff_show", ["id" => $product->getId()]));
 
@@ -89,6 +95,13 @@ class StuffController extends AbstractController
      */
     public function edit(Request $request, Product $product, Breadcrumbs $breadcrumbs): Response
     {
+        if ($this->getUser() !== $product->getUser())
+        {
+            $response = new Response();
+            $response->setStatusCode(403);
+            return $response;
+        }
+
         $breadcrumbs->addItem("Stuff", $this->get("router")->generate("stuff_index"));
         $breadcrumbs->addItem("Edit", $this->get("router")->generate("stuff_edit", ["id" => $product->getId()]));
 
@@ -112,6 +125,13 @@ class StuffController extends AbstractController
      */
     public function delete(Request $request, Product $product, Security $security): Response
     {
+        if ($this->getUser() !== $product->getUser())
+        {
+            $response = new Response();
+            $response->setStatusCode(403);
+            return $response;
+        }
+
         if ($this->isCsrfTokenValid('delete' . $product->getId(), $request->request->get('_token'))) {
             $entityManager = $this->getDoctrine()->getManager();
             $entityManager->remove($product);
